+++ src/main/java/net/meander/subtlyd/advancements/triggers/CriteriaTriggersSD.java
@@ -6,6 +6,7 @@ import net.minecraft.core.registries.BuiltInRegistries;
 
 public class CriteriaTriggersSD {
     public static final SleptInTentTrigger SLEPT_IN_TENT = register("slept_in_tent", new SleptInTentTrigger());
+    public static final StealthAttackTrigger STEALTH_ATTACK = register("stealth_attack", new StealthAttackTrigger());
 
     public static void registration() {}