+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 ## [5.0 Snapshot 1] - 6/25/26
 ### New Features
+#### Advancements
+- Added I Am Bush advancement
+  - Is awarded for performing a stealth attack from within a tall plant
 
 ### Changes
 #### Items
@@ -9,10 +12,11 @@
 - Changed Iron Grates to have the `dragon_immune` block tag
 
 ### Technical Changes
+#### Data Tags
+- Added `tall_plants` block tag
 
 #### Bug Fixes
 - Fixed bug allowing the Ender Dragon to destroy Iron Grates
 - Fixed bug causing a client/server de-sync when lighting a campfire with sticks
 - Fixed bug where unknown_server.png (commonly known as pack.png) was stretched when there's an issue with a world thumbnail
 - Fixed bug causing arrows to disappear if a stack was added onto them
--